pyproject.toml
[project]
name = "ceu-modern-data-platforms-databricks"
version = "1.0.0"
requires-python = ">=3.10,<3.14"
dependencies = [
    "databricks-sdk>=0.20.0",
]
